Move static components in own directory
