--- src/components/SmartsquareLogo.vue
@@ -1,5 +1,0 @@
-<template>
-  <div class="d-flex align-center">
-    <v-img contain src="../assets/logo.png" transition="scale-transition" width="100" />
-  </div>
-</template>
