+++ src/components/static/ErrorSnackbar.vue
