+++ src/components/static/GithubLogo.vue
@@ -7,6 +7,6 @@
     elevation="4"
     color="transparent"
   >
-    <v-img contain src="../assets/github.png" transition="scale-transition" max-width="40" />
+    <v-img contain src="../../assets/github.png" transition="scale-transition" max-width="40" />
   </v-btn>
 </template>
